Move config utility function to a seperate class
