+++ src/test/java/io/github/magiccheese1/damageindicator/TextUtilityTest.java
@@ -6,13 +6,15 @@ import org.junit.Test;
 
 import java.text.DecimalFormat;
 
+import static io.github.magiccheese1.damageindicator.config.configUtility.convertEasyHexToLegacy;
+
 public class TextUtilityTest {
 
     @Test
     public void testBasic() {
         Assert.assertEquals(
             "§x§F§F§F§F§F§F This is white",
-            Utility.convertEasyHexToLegacy("§#FFFFFF This is white")
+            convertEasyHexToLegacy("§#FFFFFF This is white")
         );
     }
 
@@ -20,7 +22,7 @@ public class TextUtilityTest {
     public void testMultiple() {
         Assert.assertEquals(
             "§x§F§F§F§F§F§F White and §x§3§3§F§F§5§7 green",
-            Utility.convertEasyHexToLegacy("§#FFFFFF White and §#33FF57 green")
+            convertEasyHexToLegacy("§#FFFFFF White and §#33FF57 green")
         );
     }
 
@@ -28,7 +30,7 @@ public class TextUtilityTest {
     public void testIgnored() {
         Assert.assertEquals(
             "§#FFFFF White with 5 and #33FF57 green without",
-            Utility.convertEasyHexToLegacy("§#FFFFF White with 5 and #33FF57 green without")
+            convertEasyHexToLegacy("§#FFFFF White with 5 and #33FF57 green without")
         );
     }
 
@@ -36,13 +38,13 @@ public class TextUtilityTest {
     public void testRealConfigValue() {
         Assert.assertEquals(
             "§x§3§3§F§F§5§7-0.#&4❤",
-            Utility.convertEasyHexToLegacy("&#33FF57-0.#&4❤")
+            convertEasyHexToLegacy("&#33FF57-0.#&4❤")
         );
     }
 
     @Test
     public void testInDecimalFormat() {
-        final String postHexConversion = Utility.convertEasyHexToLegacy("&#33FF57-0.#&4❤");
+        final String postHexConversion = convertEasyHexToLegacy("&#33FF57-0.#&4❤");
         final String postChatColourConversion = ChatColor.translateAlternateColorCodes('&', postHexConversion);
 
         final DecimalFormat decimalFormat = new DecimalFormat(postChatColourConversion);
